+++ bootstrap.php
@@ -11,7 +11,10 @@ return function (Dispatcher $events, Application $app) {
     $events->subscribe(Listeners\BazaarEnabled::class);
     $events->subscribe(Listeners\AddApiAttributes::class);
     $events->subscribe(Listeners\AddSatisConfiguration::class);
+    $events->subscribe(Listeners\SyncWasSet::class);
+    $events->subscribe(Listeners\SyncVersion::class);
 
     $app->register(Providers\ComposerEnvironmentProvider::class);
     $app->register(Providers\ExtensionProvider::class);
+    $app->register(Providers\ConsoleProvider::class);
 };