+++ js/admin/src/components/BazaarPage.js
@@ -1,9 +1,9 @@
 import Component from "flarum/Component";
-import ExtensionRepository from "flagrow/bazaar/utils/ExtensionRepository";
-import ExtensionListItem from "flagrow/bazaar/components/ExtensionListItem";
-import BazaarLoader from "flagrow/bazaar/components/BazaarLoader";
-import BazaarPageHeader from 'flagrow/bazaar/components/BazaarPageHeader';
-import CustomCheckbox from "flagrow/bazaar/components/CustomCheckbox";
+import ExtensionRepository from "./../utils/ExtensionRepository";
+import ExtensionListItem from "./ExtensionListItem";
+import BazaarLoader from "./BazaarLoader";
+import BazaarPageHeader from './BazaarPageHeader';
+import CustomCheckbox from "./CustomCheckbox";
 
 export default class BazaarPage extends Component {
     init() {
@@ -12,7 +12,7 @@ export default class BazaarPage extends Component {
         this.loading = m.prop(false);
         this.repository = m.prop(new ExtensionRepository(this.loading));
         this.repository().loadNextPage();
-        this.connected = app.data.settings['flagrow.bazaar.connected'] == 1 || false;
+        this.connected = app.data.settings['flagrow.bazaar.connected'] && app.data.settings['flagrow.bazaar.connected'] !== '0';
     }
 
     view() {
@@ -31,52 +31,84 @@ export default class BazaarPage extends Component {
     }
 
     search() {
-        return m('fieldset.ExtensionSearch', [
-            m('input[type=text].FormControl', {
-                value: this.repository().filteredBy('search'),
-                oninput: m.withAttr('value', term => {
-                    this.repository().filterBy('search', term);
-                }),
-                placeholder: app.translator.trans('flagrow-bazaar.admin.search.placeholder'),
-            }),
-            CustomCheckbox.component({
-                iconChecked: 'toggle-up',
-                state: this.repository().filterUpdateRequired(),
-                onchange: (checked) => this.repository().filterUpdateRequired(checked),
-                children: app.translator.trans('flagrow-bazaar.admin.search.filter_update_required')
-            }),
-            CustomCheckbox.component({
-                iconChecked: 'plus-square',
-                state: this.repository().filterInstalled(),
-                onchange: (checked) => this.repository().filterInstalled(checked),
-                children: app.translator.trans('flagrow-bazaar.admin.search.filter_installed')
-            }),
-            this.connected ? [
+        return m('div', [
+            m('fieldset', {className: 'ExtensionSearch'},
+                m('input[type=text].FormControl', {
+                    value: this.repository().filteredBy('search'),
+                    oninput: m.withAttr('value', term => {
+                        this.repository().filterBy('search', term);
+                    }),
+                    placeholder: app.translator.trans('flagrow-bazaar.admin.search.placeholder'),
+                })
+            ),
+            m('div', {className: 'ExtensionFilters ButtonGroup'}, [
                 CustomCheckbox.component({
-                    iconChecked: 'heart',
-                    state: this.repository().filterFavorited(),
-                    onchange: (checked) => this.repository().filterFavorited(checked),
-                    children: app.translator.trans('flagrow-bazaar.admin.search.filter_favorited')
+                    icon: 'toggle-up',
+                    className: 'Button hasIcon',
+                    state: this.repository().filterUpdateRequired(),
+                    onchange: (checked) => this.repository().filterUpdateRequired(checked),
+                    children: app.translator.trans('flagrow-bazaar.admin.search.filter_update_required')
                 }),
                 CustomCheckbox.component({
-                    iconChecked: 'shopping-cart',
-                    state: this.repository().filterSubscribed(),
-                    onchange: (checked) => this.repository().filterSubscribed(checked),
-                    children: app.translator.trans('flagrow-bazaar.admin.search.filter_subscribed')
+                    icon: 'circle-o-notch',
+                    className: 'Button hasIcon',
+                    state: this.repository().filterPending(),
+                    onchange: (checked) => this.repository().filterPending(checked),
+                    children: app.translator.trans('flagrow-bazaar.admin.search.filter_pending')
                 }),
-            ] : '',
-            CustomCheckbox.component({
-                iconChecked: 'certificate',
-                state: this.repository().filterPremium(),
-                onchange: (checked) => this.repository().filterPremium(checked),
-                children: app.translator.trans('flagrow-bazaar.admin.search.filter_premium')
-            }),
+                CustomCheckbox.component({
+                    icon: 'plus-square',
+                    className: 'Button hasIcon',
+                    state: this.repository().filterInstalled(),
+                    onchange: (checked) => this.repository().filterInstalled(checked),
+                    children: app.translator.trans('flagrow-bazaar.admin.search.filter_installed')
+                }),
+                CustomCheckbox.component({
+                    icon: 'inr',
+                    className: 'Button hasIcon',
+                    state: this.repository().filterLanguages(),
+                    onchange: (checked) => this.repository().filterLanguages(checked),
+                    children: app.translator.trans('flagrow-bazaar.admin.search.filter_languages')
+                }),
+                this.connected ? [
+                    CustomCheckbox.component({
+                        icon: 'heart',
+                        className: 'Button hasIcon',
+                        state: this.repository().filterFavorited(),
+                        onchange: (checked) => this.repository().filterFavorited(checked),
+                        children: app.translator.trans('flagrow-bazaar.admin.search.filter_favorited')
+                    }),
+                    CustomCheckbox.component({
+                        icon: 'shopping-cart',
+                        className: 'Button hasIcon',
+                        state: this.repository().filterSubscribed(),
+                        onchange: (checked) => this.repository().filterSubscribed(checked),
+                        children: app.translator.trans('flagrow-bazaar.admin.search.filter_subscribed')
+                    }),
+                ] : '',
+                CustomCheckbox.component({
+                    icon: 'certificate',
+                    className: 'Button hasIcon',
+                    state: this.repository().filterPremium(),
+                    onchange: (checked) => this.repository().filterPremium(checked),
+                    children: app.translator.trans('flagrow-bazaar.admin.search.filter_premium')
+                }),
+            ])
         ])
     }
 
     items() {
         return m('ul', {className: 'ExtensionList'}, [
             this.repository().extensions().filter(extension => {
+
+                if (this.repository().filterLanguages() && ! extension.locale()) {
+                    return false;
+                }
+
+                if (this.repository().filterPending() && ! extension.pending()) {
+                    return false;
+                }
+
                 if (this.repository().filterInstalled() && ! extension.installed()) {
                     return false;
                 }