+++ js/admin/src/utils/ExtensionRepository.js
@@ -9,10 +9,12 @@ export default class ExtensionRepository {
         this.loading = loading;
         this.resetNavigation();
         this.filterInstalled = m.prop(false);
+        this.filterPending = m.prop(false);
         this.filterUpdateRequired = m.prop(false);
         this.filterFavorited = m.prop(false);
         this.filterSubscribed = m.prop(false);
         this.filterPremium = m.prop(false);
+        this.filterLanguages = m.prop(false);
         this.filters = {
             search: '',
         };