+++ src/Listeners/AddClientAssets.php
@@ -3,7 +3,7 @@
 namespace Flagrow\Bazaar\Listeners;
 
 use DirectoryIterator;
-use Flarum\Event\ConfigureWebApp;
+use Flarum\Frontend\Event\Rendering;
 use Illuminate\Contracts\Events\Dispatcher;
 use Flarum\Event\ConfigureLocales;
 
@@ -11,16 +11,16 @@ class AddClientAssets
 {
     public function subscribe(Dispatcher $events)
     {
-        $events->listen(ConfigureWebApp::class, [$this, 'addAssets']);
+        $events->listen(Rendering::class, [$this, 'addAssets']);
         $events->listen(ConfigureLocales::class, [$this, 'addLocales']);
     }
 
-    public function addAssets(ConfigureWebApp $event)
+    public function addAssets(Rendering $event)
     {
         if ($event->isAdmin()) {
             $event->addAssets([
                 __DIR__.'/../../js/admin/dist/extension.js',
-                __DIR__.'/../../less/extension.less',
+                __DIR__.'/../../resources/less/extension.less',
             ]);
             $event->addBootstrapper('flagrow/bazaar/main');
         }
@@ -28,7 +28,7 @@ class AddClientAssets
 
     public function addLocales(ConfigureLocales $event)
     {
-        foreach (new DirectoryIterator(__DIR__.'/../../locale') as $file) {
+        foreach (new DirectoryIterator(__DIR__.'/../../resources/locale') as $file) {
             if ($file->isFile() && in_array($file->getExtension(), ['yml', 'yaml'])) {
                 $event->locales->addTranslations($file->getBasename('.'.$file->getExtension()), $file->getPathname());
             }